Implement Access Code Grant Flow
Access Code Grant Flow is implemented and can be used as:
 {ok, Identity, Response} = oauth2:issue_code_grant(
                                         ClientId,
                                         ClientSecret,
                                         RedirectUri,
                                         ClientScope)

it's up to the underlying backend to correctly assert ClientId and
ClientSecret with regards to if a client is confidential or not. Also
the implementer should verify Username and Password for the User that
the Client whishes to receive access to before issuing an Access Code.

The Response will include an access_code which is a short lived code
which can be used to retrieve an access_code as such:
{ok, Identity, Response} = oauth2:authorize_code_grant(
                                         ClientId,
                                         ClientSecret,
                                         AccessCode,
                                         RedirectUri)

The Response will include an valid Access Token and an optional Refresh
Token.