Adds response_type/grant_type parameter to API
WARNING: Breaks the compatibility with any code using the library
The specification of Authorization Code Grant and Implicit Grant
requires that unsupported_response_type errors are sent to the redirect
URI if this is valid (see 4.1.2.1 and 4.2.2.1). In order to facilitate
this, authorize_code_request and authorize_password have been modified
to return unsupported_response_type when response_type in the request is
wrong but client and redirect_uri are valid.
response_type or grant_type parameters have been added to every other
API function for the shake of consistency.
See an example of how this API could be used in

https://github.com/IvanMartinez/oauth2_webmachine/tree/test